requirements.txt
numpy>=1.26.0
pandas>=2.1.0
openpyxl==3.1.2
openai==1.35.3
httpx==0.27.2
requests==2.31.0
python-dotenv==1.0.0
fastapi==0.111.0
uvicorn[standard]==0.30.3
pydantic==2.5.0
python-multipart==0.0.9
streamlit==1.37.1
